requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langgraph==0.0.62
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
